pyproject.toml
[tool.poetry]
name = "gp-script-writer-backend"
version = "0.1.0"
description = ""
authors = ["imayabharathi-mv <imayabharathi@mindvalley.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.13.1"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"
email-validator = "^2.1.0.post1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "4.0.1"
python-multipart = "^0.0.6"
google-auth = "^2.25.2"
azure-storage-blob = "^12.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
